service-workers/service-worker/fetch-event-async-respond-with.https.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/test-helpers.sub.js"></script>
<script src="resources/service-worker-recorder.js"></script>
<script>
promise_test(function(t) {
    var script = 'resources/fetch-event-async-respond-with-worker.js';
    var scope = 'resources/simple.html';

    return service_worker_unregister_and_register(t, script, scope)
      .then(function(registration) {
          return wait_for_state(t, registration.installing, 'activated')
            .then(function() {
                return ServiceWorkerRecorder.client.clear(registration.active);
              });
        })
      .then(function() {
          return with_iframe(scope);
        })
      .then(function(frame) {
          t.add_cleanup(function() { frame.remove(); });

          return ServiceWorkerRecorder.client.read(
              frame.contentWindow.navigator.serviceWorker.controller
            );
        })
      .then(function(results) {
          assert_array_equals(results, ['PASS']);
          return service_worker_unregister_and_done(t, scope);
        })
  }, 'Calling respondWith asynchronously throws an exception');
</script>
